+++ test/main.test.ts
@@ -2,7 +2,8 @@ import { when } from "jest-when";
 
 import actionOptions from "./fixtures/svgo-action.json";
 import contentPayloads from "./fixtures/contents-payloads.json";
-import svgoOptions from "./fixtures/svgo-options.json";
+import svgoV1Options from "./fixtures/svgo-v1-options.json";
+import svgoV2Options from "./fixtures/svgo-v2-options.json";
 
 import * as core from "./mocks/@actions/core.mock";
 import * as github from "./mocks/@actions/github.mock";
@@ -111,7 +112,7 @@ test.each(ALL_EVENTS)("dry mode enabled (%s)", async (eventName) => {
   github.context.eventName = eventName;
 
   inputs.ActionConfig.mockImplementationOnce(() => {
-    return { isDryRun: true };
+    return { svgoOptionsPath: ".svgo.yml", isDryRun: true };
   });
 
   await main();
@@ -131,7 +132,7 @@ test.each(ALL_EVENTS)("use custom configuration file (%s)", async (eventName) =>
   expect(inputs.ActionConfig).toHaveBeenCalledWith(core, actionOptions);
 });
 
-test.each(ALL_EVENTS)("use an SVGO options file in the repository (%s)", async (eventName) => {
+test.each(ALL_EVENTS)("use a YAML SVGO options file in the repository (%s)", async (eventName) => {
   svgo.SVGOptimizer.mockClear();
 
   github.context.eventName = eventName;
@@ -143,7 +144,54 @@ test.each(ALL_EVENTS)("use an SVGO options file in the repository (%s)", async (
 
   await main();
 
-  expect(svgo.SVGOptimizer).toHaveBeenCalledWith(svgoOptions);
+  expect(svgo.SVGOptimizer).toHaveBeenCalledWith(1, svgoV1Options);
+});
+
+test.each(ALL_EVENTS)("use a JavaScript SVGO options file in the repository (%s)", async (eventName) => {
+  svgo.SVGOptimizer.mockClear();
+
+  github.context.eventName = eventName;
+
+  inputs.ActionConfig.mockImplementationOnce(() => {
+    return { svgoOptionsPath: ".svgo.js", svgoVersion: 1 };
+  });
+
+  when(githubAPI.getFile)
+    .calledWith(github.GitHubInstance, github.context.sha, ".svgo.js")
+    .mockResolvedValueOnce(contentPayloads.files[".svgo.js"]);
+
+  await main();
+
+  expect(svgo.SVGOptimizer).toHaveBeenCalledWith(1, svgoV2Options);
+});
+
+test("use a JavaScript SVGO options file that does not exist", async () => {
+  svgo.SVGOptimizer.mockClear();
+
+  github.context.eventName = EVENT_PUSH;
+
+  inputs.ActionConfig.mockImplementationOnce(() => {
+    return { svgoOptionsPath: ".svgo.js", svgoVersion: 1 };
+  });
+
+  when(githubAPI.getFile)
+    .calledWith(github.GitHubInstance, github.context.sha, ".svgo.js")
+    .mockRejectedValueOnce(new Error("Not found"));
+
+  await main();
+
+  expect(svgo.SVGOptimizer).toHaveBeenCalledWith(1, expect.any(Object));
+});
+
+test.each([1, 2])("set SVGO version", async (svgoVersion) => {
+  inputs.ActionConfig.mockImplementationOnce(() => {
+    return { svgoOptionsPath: ".svgo.yml", svgoVersion: svgoVersion };
+  });
+
+  await main();
+
+  expect(svgo.SVGOptimizer).toHaveBeenCalledWith(svgoVersion, expect.any(Object));
+  expect(core.info).toHaveBeenCalledWith(expect.stringMatching(`SVGO.*${svgoVersion}`));
 });
 
 test.each(ALL_EVENTS)("the Action configuration file does not exist (%s)", async (eventName) => {